refactor(search-engine): bring server/tools/search-engine.ts to the lint standard (#780)
The file carried 9 oxlint findings: 1 max-lines (837 code lines against a
standard of 500), 7 max-params, and 1 complexity. All 9 are cleared with no
behavior change -- schemas, defaults, error strings, emitted SQL, and trace
event names are identical.

Split along four seams into sibling modules, each well under the standard:

  search-engine-types.ts     row shapes, tuning constants, trace evidence
  search-engine-sql.ts       every SQL fragment the arms assemble
  search-engine-rows.ts      source_ref attachment, canonical namespaces
  search-engine-fallback.ts  the legacy-namespace merge and its dedupe

search-engine.ts keeps the embedding, the three arms, the RRF fusion, and the
dispatch between them.

The max-params findings were all one call chain threading the same values
positionally. The four CTE builders now share a CteContext, which is more than
shape: the namespace parameter index and its array-ness must agree with the
params array the arm already pushed onto, and carrying them as one resolved
value makes it impossible to pass one arm's index alongside another arm's
array-ness. The arms and executeSearchInternal take a SearchRequest.

executeSearch and executeSearchWithSharedFallback are called positionally from
server/tools/ consumers and from tests this change does not touch, so their
argument order IS the compatibility contract. They are expressed as labelled
tuple rest parameters, which keeps every position's type checked exactly as
before -- passing a Tier where mode belongs is still a compile error -- while
letting the engine carry one argument internally instead of nine or ten.

The complexity finding on executeSearchWithSharedFallback is cleared by naming
its two decisions: sharedFallbackApplies answers whether the top-up is
reachable at all, and searchWithLegacyTopUp runs it.

Reuse checked, not assumed: appendNamespaceParam/namespaceFilterSql are NOT a
private copy of read-scope.ts's appendReadNamespacePredicate. That helper takes
an AuthIdentity, resolves the readable set itself, and emits a finished
predicate against one column. The arms here bind ONE namespace value shared
across several aliased CTEs, so the push and the per-alias predicate are two
steps rather than one. The distinction is now recorded in the docblock.

The process.env read in searchEmbeddingTimeoutMs is left in place; rewiring it
belongs to its own lane.

Verified:
  oxlint --deny-warnings (5 files) -> exit 0
  bunx tsc --noEmit -> exit 0
  bun run test:isolated (24 files) -> 323 pass, 0 fail
